requirements.txt
fastapi>=0.115.0
google-cloud-secret-manager>=2.20.0
httpx>=0.27.0
itsdangerous>=2.2.0
pg8000==1.31.2
psycopg2-binary==2.9.9
pydantic>=2.7.0
python-dotenv>=1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.35
uvicorn[standard]==0.30.6